% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_square.short.R
\name{biscalew}
\alias{biscalew}
\title{biscalew :Robust M-estimates for scale.}
\usage{
biscalew(t)
}
\arguments{
\item{t}{Numerical input, usually residuals.}
}
\value{
bisquare weight for scale.
}
\description{
Tukey's bisquare family of functions.
}
