% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_square.short.R
\name{bisquare}
\alias{bisquare}
\title{bisquare function title.}
\usage{
bisquare(t, k = 4.685)
}
\arguments{
\item{t}{A number.}

\item{k}{A number.}
}
\value{
Return value.
}
\description{
bisquare function title.
}
