% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_ATAC_RNA_seq}
\alias{BRCA_ATAC_RNA_seq}
\title{RobMixReg package built-in breast cancer multi-omics data.}
\format{A data frame with 500 rows and 5 variables:
\describe{
  \item{a1}{RNA-seq vector 1}
  \item{a2}{RNA-seq vector 2}
  \item{a3}{RNA-seq vector 3}
  \item{b1}{ATAC-seq vector 1}
  \item{b2}{ATAC-seq vector 2}
  \item{b3}{ATAC-seq vector 3}
}}
\usage{
BRCA_ATAC_RNA_seq
}
\description{
A BRCA dataset from RobMixReg package. This BRCA cancer dataset has two types of sequence data inclcuding ATAC-seq and RNA-seq. The vignette example shows that using robust regression, we can identify the outlier mutually based on ATAC-seq and RNA-seq.
}
\keyword{datasets}
