% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{BRCA_source}
\alias{BRCA_source}
\title{RobMixReg package built-in BRCA source data.}
\format{A list whose length is 4:
\describe{
  \item{ATACseq_marker}{Three cell type marker genes}
  \item{RNAseq_marker}{Three cell type marker genes}
  \item{data_RNAseq_selected}{RNAseq sequencing data}
  \item{data_ATACseq_selected}{ATACseq sequencing data}
}}
\usage{
BRCA_source
}
\description{
The list which contain all the information to generate variables used in the real application.
}
\keyword{datasets}
