% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{calcobj_i}
\alias{calcobj_i}
\title{calcobj_i: it calculates the objective function}
\usage{
calcobj_i(X, Y, iter, pa)
}
\arguments{
\item{X}{independent variable.}

\item{Y}{response variable.}

\item{iter}{see original CWM method.}

\item{pa}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
function to calculate the objective function.
}
