% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{CCLE_data}
\alias{CCLE_data}
\title{RobMixReg package built-in CCLE data.}
\format{A list whose length is 2:
\describe{
  \item{X}{Gene expression dataset.}
  \item{Y}{AUCC score.}
}}
\usage{
CCLE_data
}
\description{
The list which contain all the information to generate variables used in the real application.
}
\keyword{datasets}
