% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_wrapper.R
\name{compPlot}
\alias{compPlot}
\title{The plot wrapper function.}
\usage{
compPlot(type = "rlr", x, y, nc, inds_in, res)
}
\arguments{
\item{type}{The character to choose which type of plot to generate.}

\item{x}{The independent variables}

\item{y}{The external variable}

\item{nc}{The number of components}

\item{inds_in}{A vector indicate the outlier samples.}

\item{res}{The result object returned by MLM function.}
}
\description{
The plot wrapper function.
}
