% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_SVD.R
\name{Compute_Rbase_SVD}
\alias{Compute_Rbase_SVD}
\title{Compute the row space using SVD.}
\usage{
Compute_Rbase_SVD(bulk_data, tg_R1_lists_selected)
}
\arguments{
\item{bulk_data}{The bulk data..}

\item{tg_R1_lists_selected}{A list of the marker genes for several cell types.}
}
\value{
A matrix which each row span the row space using cell type specific marker genes.
}
\description{
Compute the row space using SVD.
}
