% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBSL.R
\name{CSMR_predict}
\alias{CSMR_predict}
\title{The predict function of the CSMR algorithm.}
\usage{
CSMR_predict(CSMR_coffs, CSMR.model, xnew, ynew, singleMode = F)
}
\arguments{
\item{CSMR_coffs}{The coefficient matrix.}

\item{CSMR.model}{The trained model.}

\item{xnew}{x variable.}

\item{ynew}{y variable.}

\item{singleMode}{A parameter to set the component to one.}
}
\value{
A list object consist of coefficient, clustering membership, data x, external variable y, predicted y based on regression model.
}
\description{
The predict function of the CSMR algorithm.
}
