% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLE.short.R
\docType{methods}
\name{CTLE}
\alias{CTLE}
\alias{CTLE,formula,ANY,ANY,numeric-method}
\title{Method CTLE.}
\usage{
CTLE(formula, data, nit = 20, nc = 2)

\S4method{CTLE}{formula,ANY,ANY,numeric}(formula, data, nit = 20,
  nc = 2)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the variables in the model.}

\item{nit}{An nit parameter for CTLE method.}

\item{nc}{An optional number of clusters.}
}
\description{
Method CTLE.
}
