% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLE.short.rob.R
\docType{methods}
\name{CTLERob}
\alias{CTLERob}
\alias{CTLERob,formula,ANY,ANY,numeric-method}
\alias{CTLERob,formula,ANY,ANY,numeric,ANY-method}
\title{CTLERob: Robust mixture regression based on component-wise adaptive trimming likelihood estimation.}
\usage{
CTLERob(formula, data, nit = 20, nc = 2, rlr_method = "ltsReg")

\S4method{CTLERob}{formula,ANY,ANY,numeric}(formula, data, nit = 20,
  nc = 2, rlr_method = "ltsReg")
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the predictor and response variables, where the last column is the response varible.}

\item{nit}{Number of iterations.}

\item{nc}{Number of mixture components.}

\item{rlr_method}{The regression methods, default is 'ltsReg'.}
}
\description{
CTLERob performes robust linear regression with high breakdown point and high efficiency in each mixing components and adaptively remove the outlier samples.
}
