% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixL.short.R
\name{denLp}
\alias{denLp}
\title{denLp : Density function for Laplace distribution.}
\usage{
denLp(rr, sig)
}
\arguments{
\item{rr}{Shift from the location parameter}

\item{sig}{Scale parameter.}
}
\value{
Laplace density.
}
\description{
Laplace distribution.
}
