% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLE.short.R
\name{DeOut}
\alias{DeOut}
\title{DeOut : Detect outlier observations.}
\usage{
DeOut(daData, method)
}
\arguments{
\item{daData}{A numerical vector.}

\item{method}{Choose from '3sigma','hampel' and 'boxplot'.}
}
\value{
indices of outlier observations.
}
\description{
Detect outlier observations from a vector.
}
