% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{dmnorm}
\alias{dmnorm}
\title{dmnorm: multivariate normal density function}
\usage{
dmnorm(X, mu, sigma)
}
\arguments{
\item{X}{see original CWM method.}

\item{mu}{see original CWM method.}

\item{sigma}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
