% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{estimClustPar_i}
\alias{estimClustPar_i}
\title{estimClustPar_i: see original CWM method}
\usage{
estimClustPar_i(X, Y, iter, pa)
}
\arguments{
\item{X}{independent variables.}

\item{Y}{response variable.}

\item{iter}{see original CWM method.}

\item{pa}{see original CWM method.}
}
\value{
xsee original CWM method.
}
\description{
see original CWM method.
}
