% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{findClustAssig_i}
\alias{findClustAssig_i}
\title{findClustAssig_i: it obtains the assigment corresponding to each point to the K populations and the set of trimmed points}
\usage{
findClustAssig_i(X, Y, iter, pa)
}
\arguments{
\item{X}{independent variable.}

\item{Y}{response variable.}

\item{iter}{see original CWM method.}

\item{pa}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
