% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{getini}
\alias{getini}
\title{getini: see original CWM method}
\usage{
getini(K, n)
}
\arguments{
\item{K}{see original CWM method.}

\item{n}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
