% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{InitClusters_i}
\alias{InitClusters_i}
\title{InitClusters_i: see original CWM method}
\usage{
InitClusters_i(X, Y, iter, pa)
}
\arguments{
\item{X}{independent variable.}

\item{Y}{response variable.}

\item{iter}{see original CWM method.}

\item{pa}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
