% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_square.short.R
\docType{methods}
\name{mixlinrb_bi}
\alias{mixlinrb_bi}
\alias{mixlinrb_bi,formula,ANY,numeric,numeric-method}
\title{mixlinrb_bi: mixlinrb_bione estimates the mixture regression parameters robustly using bisquare function based on multiply initial value.}
\usage{
mixlinrb_bi(formula, data, nc = 2, nit = 200)

\S4method{mixlinrb_bi}{formula,ANY,numeric,numeric}(formula, data,
  nc = 2, nit = 20)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the predictor and response variables, where the last column is the response varible.}

\item{nc}{Number of mixture components.}

\item{nit}{Number of iterations for biSauqre method.}
}
\value{
Estimated coefficients of all components.
}
\description{
An EM-type of parameter estimation by replacing the least square estimation in the M-step with a robust criterior.
}
