% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_square.short.R
\name{mixlinrb_bione}
\alias{mixlinrb_bione}
\title{mixlinrb_bione function title.}
\usage{
mixlinrb_bione(formula, data, nc = 2)
}
\arguments{
\item{formula}{A formula.}

\item{data}{A parameter.}

\item{nc}{A parameter.}
}
\value{
Return value.
}
\description{
mixlinrb_bione function title.
}
