% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixL.short.R
\docType{methods}
\name{mixLp}
\alias{mixLp}
\alias{mixLp,formula,ANY,numeric,numeric-method}
\alias{CTLE,formula,ANY,numeric,numeric-method}
\title{Method mixLp.}
\usage{
mixLp(formula, data, nc=2, numini=200)

\S4method{mixLp}{formula,ANY,numeric,numeric}(formula, data, nc = 2,
  numini = 20)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the variables in the model.}

\item{nc}{An optional number of clusters.}

\item{numini}{The minimal value of nu.}
}
\value{
xxxxxx
}
\description{
mixLp estimates the mixture regression parameters robustly using bisquare function based on multiple initial values. The solution is found by the modal solution.
}
