% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mixL.short.R
\name{mixLp_one}
\alias{mixLp_one}
\title{mixLp_one function.}
\usage{
mixLp_one(formula, data, nc = 2)
}
\arguments{
\item{formula}{The parameter.}

\item{data}{The parameter.}

\item{nc}{The number of component.Default value is 2.}
}
\value{
Return value.
}
\description{
The robust EM algorithm to fit the mixture of linear regression based on bisquare function. Bai, X., Yao, W._, and Boyer, J. E. (2012). Robust fitting of mixture regression models.
}
