% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vali_mode.R
\name{MLM_bic}
\alias{MLM_bic}
\title{Model selection function for low dimension data.}
\usage{
MLM_bic(ml.method = "rlr", x, y, nc = 1, formulaList = NULL, K = 2)
}
\arguments{
\item{ml.method}{The parameter to choose the fitted model for calculating the BIC}

\item{x}{x variable.}

\item{y}{y variable.}

\item{nc}{The component number for low dimensional feature}

\item{formulaList}{The list of target formular}

\item{K}{The component number for high dimensional feature}
}
\value{
BIC value.
}
\description{
Model selection function for low dimension data.
}
