% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vali_mode.R
\name{MLM_cv}
\alias{MLM_cv}
\title{Cross validation (fold-5) function for high dimension data.}
\usage{
MLM_cv(x = NULL, y = NULL, nit = 1, nc = 2, max_iter = 50)
}
\arguments{
\item{x}{x variable.}

\item{y}{y variable.}

\item{nit}{Iteration number.}

\item{nc}{The number of component.}

\item{max_iter}{Maximum iteration.}
}
\value{
The correlation between y and y_hat based on five fold cross validation.
}
\description{
Cross validation (fold-5) function for high dimension data.
}
