% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_RobMixReg.R
\docType{methods}
\name{plot_CTLE}
\alias{plot_CTLE}
\alias{plot_CTLE,formula,ANY,numeric-method}
\alias{CTLERob,formula,ANY,numeric,ANY}
\title{plot_CTLE: Plot the mixture/single regression line(s) in a simply function.}
\usage{
plot_CTLE(formula, data, nc = 2, inds_in)

\S4method{plot_CTLE}{formula,ANY,numeric}(formula, data, nc = 2, inds_in)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the predictor and response variables, where the last column is the response varible.}

\item{nc}{Number of mixture components.}

\item{inds_in}{The index of the point which belongs to the current regression line.}
}
\description{
CTLERob performes robust linear regression with high breakdown point and high efficiency in each mixing components and adaptively remove the outlier samples.
}
