% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{restr.diffax_i2}
\alias{restr.diffax_i2}
\title{restr.diffax_i2: see original CWM method}
\usage{
restr.diffax_i2(iter, pa)
}
\arguments{
\item{iter}{see original CWM method.}

\item{pa}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
