% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{restr2_eigenv}
\alias{restr2_eigenv}
\title{restr2_eigenv: see original CWM method}
\usage{
restr2_eigenv(autovalues, ni.ini, factor, zero.tol)
}
\arguments{
\item{autovalues}{see original CWM method.}

\item{ni.ini}{see original CWM method.}

\item{factor}{see original CWM method.}

\item{zero.tol}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
