% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLE.short.R
\docType{class}
\name{RobMixReg-class}
\alias{RobMixReg-class}
\title{Class RobMixReg.}
\description{
Class \code{RobMixReg} defines a robust mixture regression class as a S4 object.
}
\section{Slots}{

\describe{
\item{\code{inds_in}}{The index number in the regression predicted line.(non-outlier)}

\item{\code{indout}}{The index number not in the regression predicted line.(outlier)}

\item{\code{ctleclusters}}{The cluster information.}

\item{\code{compcoef}}{Component coefficient.}

\item{\code{comppvals}}{Component p values.}

\item{\code{call}}{Call function.}
}}

