% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ROSMAP_data}
\alias{ROSMAP_data}
\title{RobMixReg package built-in ROSMAP data.}
\format{A list whose length is 2:
\describe{
  \item{clincal_data}{Clinical information of ROSMAP dataset.}
  \item{rna}{The gene expression value matrix.}
}}
\usage{
ROSMAP_data
}
\description{
The list which contain all the information to generate variables used in the real application.
}
\keyword{datasets}
