% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RBSL.R
\name{simu_data_sparse}
\alias{simu_data_sparse}
\title{Simulate high dimension data for RBSL algorithm validation.}
\usage{
simu_data_sparse(n, bet, pr, sigma)
}
\arguments{
\item{n}{Patient number.}

\item{bet}{The coefficient matrix.}

\item{pr}{A vector of probability threshold which simulate the sampling based on uniform distribution.}

\item{sigma}{A vector of noise level. The length should be equal to the component number.}
}
\value{
A list object consist of x, y, true cluster label.
}
\description{
Simulate high dimension data for RBSL algorithm validation.
}
