% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_func}
\alias{simu_func}
\title{The simulation function for low/high dimensional space.}
\usage{
simu_func(beta, sigma, alpha = NULL, n = 400)
}
\arguments{
\item{beta}{The slope vector for low dimensional space or matrix for high dimensional space.}

\item{sigma}{A vector whose k-th element is the standard deviation for the k-th regression component.}

\item{alpha}{The parameter to control the number of outliers for low dimensional space.}

\item{n}{The sample number for high dimensional data.}
}
\value{
A list object.
}
\description{
The simulation function for low/high dimensional space.
}
