% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulation.R
\name{simu_low}
\alias{simu_low}
\title{The simulation function for low dimensional space.}
\usage{
simu_low(beta, inter, alpha = NULL)
}
\arguments{
\item{beta}{The slope vector.}

\item{inter}{The intercept vector.}

\item{alpha}{The parameter to control the number of outliers.}
}
\value{
A list object consists of the x variable in low dimensional space and the external y variable.
}
\description{
The simulation function for low dimensional space.
}
