% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{simuData}
\alias{simuData}
\title{RobMixReg package built-in simulated example data.}
\format{A data frame with 500 rows and 5 variables:
\describe{
  \item{X1}{X1 variable}
  \item{X2}{X2 variable}
  \item{y}{y variable}
  \item{c}{cluster information}
  \item{outlier}{outlier indicator}
}}
\usage{
simuData
}
\description{
A simulation dataset from RobMixReg package. This simulation dataset is in dimension 2 and ground truth (include outliers label) of the cluster information also generated.
}
\keyword{datasets}
