% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{ssclmat}
\alias{ssclmat}
\title{ssclmat: see original CWM method}
\usage{
ssclmat(x, k)
}
\arguments{
\item{x}{see original CWM method.}

\item{k}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
