% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TLE.short.R
\docType{methods}
\name{TLE}
\alias{TLE}
\alias{TLE,formula,ANY,numeric,numeric,numeric-method}
\alias{CTLE,formula,ANY,numeric,numeric,numeric-method}
\title{Method TLE.}
\usage{
TLE(formula, data, nc = 2, tRatio, MaxIt = 200)

\S4method{TLE}{formula,ANY,numeric,numeric,numeric}(formula, data,
  nc = 2, tRatio, MaxIt = 200)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the variables in the model.}

\item{nc}{An optional number of clusters.}

\item{tRatio}{An tRatio parameter for TLE method.}

\item{MaxIt}{An MaxIt parameter for TLE method.}
}
\value{
xxx
}
\description{
Method TLE.
}
