% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{TreatSingularity}
\alias{TreatSingularity}
\title{TreatSingularity: see original CWM method}
\usage{
TreatSingularity(iter, pa)
}
\arguments{
\item{iter}{see original CWM method.}

\item{pa}{see original CWM method.}
}
\value{
see original CWM method.
}
\description{
see original CWM method.
}
