% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CWM.short.R
\name{trim.cwm}
\alias{trim.cwm}
\title{trim.cwm: main function}
\usage{
trim.cwm(
  X,
  Y,
  K,
  niter = 20,
  Ksteps = 10,
  maxfactx = 5,
  maxfacty = 5,
  zero.tol = 1e-16,
  zero.tol1 = 1e-16,
  zero.tol2 = 1e-90,
  trace = 0,
  alpha = 0.1,
  init = 5
)
}
\arguments{
\item{X}{independent variables.}

\item{Y}{response variable.}

\item{K}{number of clusters.}

\item{niter}{see original CWM method.}

\item{Ksteps}{see original CWM method}

\item{maxfactx}{see original CWM method}

\item{maxfacty}{see original CWM method}

\item{zero.tol}{see original CWM method}

\item{zero.tol1}{see original CWM method}

\item{zero.tol2}{see original CWM method}

\item{trace}{see original CWM method}

\item{alpha}{see original CWM method}

\item{init}{number of initializations}
}
\value{
an RobMixReg object.
}
\description{
main function adapted from the original developer of CWM method.
}
