% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vali_mode.R
\name{vali_model}
\alias{vali_model}
\title{Model selection function for low dimension data.}
\usage{
vali_model(outcome, gene)
}
\arguments{
\item{outcome}{y variable.}

\item{gene}{x variable.}
}
\description{
Model selection function for low dimension data.
}
