% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_square.short.R
\name{bisquare}
\alias{bisquare}
\title{bisquare : Robust estimates for mean.}
\usage{
bisquare(t, k = 4.685)
}
\arguments{
\item{t}{Numerical input, usually residuals.}

\item{k}{A constant tuning parameter, default is 4.685.}
}
\value{
A bi-square weight for mean.
}
\description{
Tukey's bisquare family of functions.
}
