% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLE.short.R
\name{DeOut}
\alias{DeOut}
\title{DeOut function title.}
\usage{
DeOut(daData, method)
}
\arguments{
\item{daData}{A parameter.}

\item{method}{Choose from '3sigma','hampel' and 'boxplot'.}
}
\value{
The result explanation.
}
\description{
DeOut function title.
}
