% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLE.short.R
\name{flexmix_2}
\alias{flexmix_2}
\title{flexmix_2 function title.}
\usage{
flexmix_2(formula, data1, k, mprior)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit. The general form is y~x|g where y is the response, x the set of predictors and g an optional grouping factor for repeated measurements.}

\item{data1}{The data matrix for mixture regression.}

\item{k}{The default mixture regression line number.}

\item{mprior}{A number.}
}
\value{
The result explanation.
}
\description{
flexmix_2 function title.
}
