% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{gaussData}
\alias{gaussData}
\title{RobMixReg package built-in gaussian example data.}
\format{A data frame with 100 rows and 3 variables:
\describe{
  \item{x}{x variable}
  \item{y}{y variable}
  \item{c}{cluster information}
}}
\usage{
gaussData
}
\description{
A dataset generated from gaussian distribution in RobMixReg package.
}
\keyword{datasets}
