% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_square.short.R
\docType{methods}
\name{mixlinrb_bi}
\alias{mixlinrb_bi}
\alias{mixlinrb_bi,formula,ANY,numeric,numeric-method}
\title{Method mixlinrb_bi.}
\usage{
mixlinrb_bi(formula, data, nc = 2, numini = 200)

\S4method{mixlinrb_bi}{formula,ANY,numeric,numeric}(formula, data,
  nc = 2, numini = 20)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the variables in the model.}

\item{nc}{An optional number of clusters.}

\item{numini}{An numini parameter for biSauqre method.}
}
\description{
Method mixlinrb_bi.
}
