% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bi_square.short.R
\name{mixlinrb_bione}
\alias{mixlinrb_bione}
\title{mixlinrb_bione : mixlinrb_bione estimates the mixture regression parameters robustly using bisquare function based on one initial value.}
\usage{
mixlinrb_bione(formula, data, nc = 2)
}
\arguments{
\item{formula}{A symbolic description of the model to be fit.}

\item{data}{A data frame containing the predictor and response variables, where the last column is the response varible.}

\item{nc}{Number of mixture components.}
}
\value{
Estimated coefficients of all components.
}
\description{
An EM-type of parameter estimation by replacing the least square estimation in the M-step with a robust criterior.
}
