% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outer.R
\name{rmr}
\alias{rmr}
\title{Robust Mixture Regression using four methods.}
\usage{
rmr(lr.method = "flexmix", formula = NULL, data = NULL, nc = 2,
  nit = 20, tRatio = 0.05, MaxIt = 200, numini = 20)
}
\arguments{
\item{lr.method}{A robust mixture regression method to be used. Should be one of "flexmix","TLE","CTLE","biSquare".}

\item{formula}{A symbolic description of the model to be fit. The general form is y~x|g where y is the response, x the set of predictors and g an optional grouping factor for repeated measurements.}

\item{data}{A data frame containing the variables in the model.}

\item{nc}{An optional number of clusters.}

\item{nit}{An nit parameter for CTLE method.}

\item{tRatio}{An tRatio parameter for TLE method.}

\item{MaxIt}{An MaxIt parameter for TLE method.}

\item{numini}{An numini parameter for biSauqre method.}
}
\value{
An S4 object about the regression result.
}
\description{
Robust Mixture Regression using four methods.
}
\examples{
\dontrun{
library(RobMixReg)
library(robust)
library(flexmix)
library(robustbase)
library(MASS)
library(gtools)
# gaussData
x=(gaussData$x);y=as.numeric(gaussData$y);
formula01=as.formula("y~x")
example_data01=data.frame(x,y)
res_rmr = rmr(lr.method='flexmix', formula=formula01, data=example_data01)
res_rmr = rmr(lr.method='CTLE', formula=formula01, data=example_data01)
}
}
