% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CTLE.short.R
\docType{class}
\name{RobMixReg-class}
\alias{RobMixReg-class}
\title{Class RobMixReg.}
\description{
Class \code{RobMixReg} defines a robust mixture regression class as a S4 object.
}
\section{Slots}{

\describe{
\item{\code{inds_in}}{The indices of observations used in the parameter estimation.}

\item{\code{indout}}{The indices of outlier samples, not used in the parameter estimation.}

\item{\code{ctleclusters}}{The cluster membership of each observation.}

\item{\code{compcoef}}{Regression coefficients for each component.}

\item{\code{comppvals}}{Component p values.????}

\item{\code{call}}{Call function.}
}}

